MYSQL-15: Implement server-side statement timeout support
Allow threads waiting inside a storage engine to be interrupted if
the connection or statement is terminated. This is a prerequisite
for implementing server-side statement timeout support with
granularity of milliseconds. For example, due to the way that the
InnoDB background lock timeout thread works, interrupted transactions
are only detected once every second.