MYSQL-15: Implement server-side statement timeout support
Introduce a server-side time limit for the execution of statements.
This feature works by interrupting the execution of statements that
take over a specified number of milliseconds to complete. After the
specified number of milliseconds has passed, the server attempts to
abort the statement without affecting the session (connection)
itself.

The implementation is optimized for the case where a statement
finishes executing before reaching the time limit. Also, because the
timeout event is an asynchronous operation to the statement being
executed, it is possible that a given statement may not be
interrupted immediatelyafter the time limit is exceeded.

The maximum execution time can be set with the max_statement_time
variable. Using the @@max_statement_time variable a user should be
able to specify the maximum execution time for any statement. The
variable value is based on a unsigned 64 bits long integer, which
represents milliseconds. An infinite timeout is represented by a
zero value, meaning that no timeout value is set.

Lastly, the maximum execution time only applies to top-level
statements or queries; compound statements are treated as a regular
component of the top-level statement.