MYSQL-15: Implement server-side statement timeout support
Implement an option in the query specification to set the maximum
execution time for a SELECT statement. Using the MAX_STATEMENT_TIME=N
option, a user is able to specify the maximum execution time (in
milliseconds) for a SELECT statement. For example, SELECT
MAX_STATEMENT_TIME=10 * FROM table.

The source value (N) can only be an integer literal. The option
is intended mainly for queries and is not supported if the SELECT
statement where the option is used is not a top-level statement
(e.g. a subquery). Also, the MAX_STATEMENT_TIME option takes
precedence over @@max_statement_time.