MYSQL-15: Implement server-side statement timeout support
Introduce an account resource limit named MAX_STATEMENT_TIME that can
be used to set the default maximum execution time for any statement
issued under a given account. That is, the max_statement_time session
variable is initialized at connect time using the current value of the
equally named resource limit.

The MAX_STATEMENT_TIME limit can be set with a CREATE USER or GRANT
statement. For example, GRANT ... TO 'user'@'host' WITH
MAX_STATEMENT_TIME 10. The server stores the specified value for an
account in the max_statement_time column of the user table row
corresponding to the account.

Given that the max_statement_time variable is initialized using the
limit value, the global counterpart of the variable is removed.