Bug#13825266 RACE IN LOCK_VALIDATE() WHEN ACCESSING PAGES DIRECTLY FROM BUFFER POOL
rb://975
approved by: Marko Makela

There is a race in lock_validate() where we try to access a page
without ensuring that the tablespace stays valid during the operation
i.e.: it is not deleted. This patch tries to fix that by using an
existing flag (the flag is renamed to make it's name more generic
in line with it's new use).

--BZR--
revision-id: inaam.rana@oracle.com-20120315173017-f66kqnu2dshfe6ii
property-branch-nick: mysql-5.1
testament3-sha1: e53270f2cf0609c57fe762ee3969c8f112e669c3