Use the Google Breakpad crash reporting system to generate minimal crash dump information when the server process (mysqld) crashes.
The minidump file generated by breakpad contains a list of the
executable and shared libraries loaded in the process, the state
of the processor register and a stack trace for each thread, and
miscellaneous information about the system and the reason for the
dump. Minidumps are significantly smaller than core files, making
them more practical for collection and processing.

The option minidump_dir can be used to specify the path name of
the directory in which to store minidumps. Also, a minidump can
be generated at runtime by requesting the server to dump debug
information with the mysqladmin debug command.