Bug#65061: LRU flush rate calculation is based on invalid values
The internal counter for the number of pages flushed as part of LRU
flushing is being incremented twice, once in buf_flush_LRU_list_batch()
and again in buf_flush_common().

This might lead to an incorrect calculation of the rate at which LRU
flush is happening, which might later affect the statistics associated
with adaptive flushing.