Bug#65310: Crash with partition pruning and impossible condition
Reading from a partitioned table using a impossible (e.g. where,
between, etc) condition leads to a crash if all partitions are pruned
and the optimizer attempts to obtain the number of rows in the table.

The problem occurs because even though no partitions are used, the
optimizer might still attempt to perform (e.g. range) analysis where an
estimate of the the number of rows is calculated using records_in_range,
which the partition engine wasn't prepared to handle.

The solution is to use zero as the estimated number of rows in the
table if no partition is being used.