Bug #12752572 61579: REPLICATION FAILURE WHILE INNODB_AUTOINC_LOCK_MODE=1 AND USING TRIGGER
When an insert stmt like "insert into t values (1),(2),(3)" is
executed, the autoincrement values assigned to these three rows are
expected to be contiguous.  In the given lock mode
(innodb_autoinc_lock_mode=1), the auto inc lock will be released
before the end of the statement.  So to make the autoincrement
contiguous for a given statement, we need to reserve the auto inc
values at the beginning of the statement.  

Modified the fix based on review comment by Svoj.  

--BZR--
revision-id: annamalai.gurusami@oracle.com-20120521115540-xd2d6sqn4gegehcv
property-audit-revid: annamalai.gurusami@oracle.com-20120521115540-9l6w5cm4c4vmcfc1
property-branch-nick: mysql-5.1
testament3-sha1: 29655b4690df288ba297749bdd08905890ca6a17