Bug#65469: Infinite loop when opening a corrupted table
If a corrupted page is read when calculating statistics for a table,
InnoDB will forever keep trying to read the page.

If the page cannot be read after a fixed number of attempts, abort
the server. The table might be accessed later by forcing InnoDB
recovery (e.g. innodb_force_recovery = 6).