Improve obscure "Got error -1 when reading table" messages
The problem is that handler originated errors printed to the error log
sometimes do not indicate the precise nature of the problem. This is
especially an issue with InnoDB because it cannot always convert an
internal InnoDB error code to a MySQL error code.

The solution is to change InnoDB so that it provides an additional
error message if it cannot convert the error code. Additionally,
errors related to a statement being interrupted are no longer printed
to the error log.

The error message format is also changed from "Got error when reading
table" to "Error when reading table: error message (error number)" in
order to distinguish the offending error message.