Bug#65715: Wrong connection ID (thread ID) in the general and slow query logs
The problem is that connection IDs (thread IDs) above 2^32 are not
properly written to the general query log. The method used to write
commands to the log (e.g. MYSQL_QUERY_LOG::write) uses a 32-bits wide
integer to store the connection ID value, which is actually a 64-bits
wide integer value, thus causing the higher bits of the value to be
discarded.

The solution is to simply use the correct type for storing thread IDs,
which is my_thread_id (unsigned long).