Error when creating RBR-safe non-deterministic stored function due to binlog enabled.
In order to prevent users from creating stored functions that will
cause unsafe events to be written to the binary log, creating a
non-deterministic function fails when binary logging is enabled. The
problem is that this applies even if the binary log format is set to
row-based, which does not have the unsafe events issue.

Even though the creation of non-deterministic functions can be allowed
by enabling log_bin_trust_function_creators, it implies that unsafe
functions won't fail at runtime if the binary log format is set to
statement. Hence, it makes more sense to allow non-deterministic
functions to be created when the binary log format is set row-based.
If said functions are invoked at runtime when the binary log format is
statement-based, they will fail with an error.