Bug#65830: utf8_general_cs is not available
The problem is that although the utf8_general_cs collation is enabled
in CMake, the config option (HAVE_UTF8_GENERAL_CS) is not being copied
over to the configuration header (e.g. config.h) because it lacks a
stub in config.h.cmake for CMake to replace with the actual value of
the variable.