Bug #13115401: -SSL-KEY VALUE IS NOT VALIDATED AND IT ALLOWS INSECURE                CONNECTIONS IF SPE
Problem description: -ssl-key value is not validated, you can assign any bogus 
text to --ssl-key and it is not verified that it exists, and more importantly, 
it allows the client to connect to mysqld.

Fix: Added proper validations checks for --ssl-key.

Note:
1) Documentation changes require for 5.1, 5.5, 5.6 and trunk in the sections
   listed below and the details are :

 http://dev.mysql.com/doc/refman/5.6/en/ssl-options.html#option_general_ssl
    and
 REQUIRE SSL section of
 http://dev.mysql.com/doc/refman/5.6/en/grant.html

2) Client having with option '--ssl', should able to get ssl connection. This 
will be implemented as part of separate fix in 5.6 and trunk.
--BZR--
revision-id: venkata.sidagam@oracle.com-20120811101304-pbyrsj1skqxolux9
property-audit-revid: venkata.sidagam@oracle.com-20120811101304-2s4zwfcmeqe722j8
property-branch-nick: 5.1
testament3-sha1: 3216e13da8911138f93adb73b759287cfd966e69