Bug#13417440 : 63340: ARCHIVE FILE IO NOT INSTRUMENTED
WARNING

This patch is for mysql-5.5 only,
to be null-merged to mysql-5.6 and mysql-trunk.

This is a partial rollback of the file io instrumentation,
removing the instrumentation for mysql_file_stat in the archive engine.

See the bug comments for details.

--BZR--
revision-id: marc.alff@oracle.com-20120824080159-je95dktj4wr7sh22
property-audit-revid: marc.alff@oracle.com-20120824080159-2hxr6zhy09hkcju9
property-branch-nick: mysql-5.5-cleanup
testament3-sha1: f7d9ab0a1fd5aa26f311c4533afa62370ba04597