Bug #14181049: MYSQL_INSTALL_DB.PL CREATES EMPTY SYSTEM TABLES FOR MYSQL
The script is different from what's used on unixes.
It was not playing the table insertion script (mysql_system_tables_data.sql),
although it was checking for the presence of this script.
Fixed by re-enabling the lookup for this file and replaying it at bootstrap
time. Note that on the Unixes "SELECT @@hostname" does return a fully qualified
name, whereas on Windows it returns only a hostname.
So by default we're filtering records in the mysql.user table until we ensure
this is fixed.

--BZR--
revision-id: georgi.kodinov@oracle.com-20120824120131-gycsz61moafhak87
property-audit-revid: georgi.kodinov@oracle.com-20120824120131-dzc0kecmgn9w957k
property-branch-nick: B14181049-5.5
property-rebase-of: georgi.kodinov@oracle.com-20120824120131-a1ohrbwjqex4a85g
testament3-sha1: fbaeb1442a06fbfc12c927f0e54f0da9651b7318