Bug#14594600 ASSERT FROM DROP TABLE CONCURRENT WITH IBUF MERGES
rb://1293
approved by: Marko Makela

There is race when dropping a single table tablespace where a reader
thread can initiate a read request before the delete flag is set and
before it is finished the deleting thread can attempt to free the
fil_node.

This patch checks the status in fil_io() to make sure that the
tablespace is not being deleted. If it is being deleted then
an error is returned instead of attempting IO.

--BZR--
revision-id: inaam.rana@oracle.com-20120920124433-ilebuy7y9cv326z9
property-audit-revid: inaam.rana@oracle.com-20120920124433-ze3fztn1d70vl5u7
property-branch-nick: mysql-5.5
testament3-sha1: 6681054dd066ce7598a536aff472f8e6a4a3fe7a