Bug#14621627 THREAD CACHE IS UNFAIR
When a client connects to a MySQL server, first a THD object is created.
If there are any idle server threads waiting, the THD object is then added
to a list and a server thread is woken up. This thread then retrieves the 
THD object from the list and starts executing.

The problem was that this list of THD objects waiting for a server thread,
was not working in a FIFO fashion, but rather LIFO. This is unfair, as it means
that the last THD added (=last client connected) will be assigned a  server 
thread first.

Note however that for this to be a problem, several clients must be able
to connect and have THD objects constructed before any server threads
manages to be woken up. This is not a very likely scenario.

This patch fixes the problem by changing the THD list to work FIFO
rather than LIFO.

This is the 5.1/5.5 version of the patch.
--BZR--
revision-id: jon.hauglid@oracle.com-20120925110953-dp6c5jyqxm90e3y8
property-audit-revid: jon.hauglid@oracle.com-20120925110953-mhdt61kqqnezkcmt
property-branch-nick: mysql-5.1-test
testament3-sha1: 0849e59404d6d398aa930837611a012956dd4c79